Implements nMedium for MieQCoreShell
I am not 100% sure it is correct, but it seems to be. This just implements the choice of the surrounding medium for the coreshell calculation and is basically what was done for MieQ(). 
The modification was based on ref: Quirantes, A., & Delgado, A. V. (1997). The scattering of light by a suspension of coated spherical particles: effects of polydispersity on cross sections. Journal of Physics D: Applied Physics, 30(15), 2123–2131. doi:10.1088/0022-3727/30/15/002 

I've compared the results with other online mie calculators (https://nanocomposix.com/pages/mie-theory-calculator) and (https://saviot.cnrs.fr/miecoat/index.en.html) which seem to show similar* results (*only discrepancies with the choice of the dielectric functions used).
However, the way I've modified the code means that xCore and xShell are redefined a second time which is not ideal but it works.